requirements.txt
SpeechRecognition
pyautogui
pynput
PyAudio
openai-whisper
torch
numpy